Add files via upload
Victor's code revisited. The file test.py is the one to run. It should allow for the evaluation of calibration as well as model calibration. The model that has been modified so far is in the file mixture_model2.py. The initial version is named mixture_model.py. In my version, the loss includes a regularization term. For now, the code doesn't work correctly since the regularization term returned is -inf.